requirements.txt
fastapi>=0.124.0
uvicorn~=0.38.0
sqlalchemy>=2.0.29
psycopg>=3.2.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
passlib>=1.7.4
email-validator>=2.1.0
psycopg2-binary>=2.9.0

